Move image components from Qt extension
